2 new benchmarks, and better checking of results
This changes the applications so that instead of having every
request return the path, each request should return a hard coded
string unique to the path.  This ensures that the benchmark is
actually working when the CHECK environment variable is used.

This adds 2 new benchmarks, one for cuba and one for roda with
the multi_route plugin.

You can now test only specific applications by specifying the
R10K_APPS environment variable, which should be a whitespace
separated list of apps to test.