Switch to benchmarking dynamic routes
Keep old static route benchmarks, which can now be run with
static_bench and static_graph rake tasks.

Add check task for checking for correct responses instead of
benchmarking. Previously, you had to manually provide a
CHECK environment variable, and it still did the same
benchmarking and data file writing.  Limit checking to 2
routes per level to save time, and add additional checks
to make sure incomplete routes and incorrect request methods
are handled correctly.  I'd like to enforce correct handling
of trailing slashes, but only Roda and Sinatra handle those
correctly, so leave that as a warning instead of a failure.

Adjust the frameworks tested by default, using a cutoff based
on the number of gem downloads.  However, because testing
Sinatra takes so long, do not test it by default.

Modify the benchmarking to avoid as much overhead as possible
when benchmarking, by precalculating all of the environment
hashes and expected return values, so the benchmarking as
much as possible only shows time spent in the application
being benchmarked.