Use `ShowExceptions` to handle 404s
This aligns the Rails benchmark more closely with the default
configuration of applications.

For the "why", `*unmatched` hits a pathological case in Rails' router
where it ends up doing lots of `match?`ing and `slice`ing due to the
path being a Regexp that can cross the `/.?` boundaries that the Router
splits strings on. Since the "more realistic" configuration avoids this
pathological case, it seems like a good change to make.